Extract choropleth hook. Dont select not selectable geometry. Fix missing node chrash
